pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "hatchling"]
build-backend = "hatchling.build"


[project]
name = "DeepACSA"
version = "0.3.1"
authors = [
  { name="Paul Ritsche", email="paul.ritsche@unibas.ch" },
]
description = "Automatic analysis of transversal muscle ultrasonography images"
readme = "README_Pypi.md"
requires-python = ">=3.9"
classifiers = [
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "Keras==2.9.0",
    "matplotlib==3.5.2",
    "numpy==1.21.2",
    "opencv-contrib-python==4.5.3.56",
    "pandas==1.3.3",
    "Pillow==8.3.2",
    "scikit-image==0.18.3",
    "scikit-learn==0.24.2",
    "tensorflow==2.9.0",
    "tqdm==4.62.2",
    "openpyxl==3.0.9",
    "h5py==3.4.0",
]

[project.urls]
"Homepage" = "https://github.com/PaulRitsche/DeepACSA"
"Bug Tracker" = "https://github.com/PaulRitsche/DeepACSA/issues"
